修改DTO属性定义错误，枚举改为int

+++ Application.PurchaseBC.Query/TradeQueries/OrderQuery.cs
@@ -341,7 +341,7 @@ namespace UniCloud.Application.PurchaseBC.Query.TradeQueries
                             EstimateDeliveryDate = l.EstimateDeliveryDate,
                             Note = l.Note,
                             BFEMaterialId = l.BFEMaterialId,
-                            Status = l.Status
+                            Status = (int) l.Status
                         }).ToList(),
                     ContractContents = o.ContractContents.Select(c => new ContractContentDTO
                     {
